adds PUT linkDeviceToEmployee

+++ src/main/java/com/device_manager/service/EmployeeService.java
@@ -1,17 +1,17 @@
 package com.device_manager.service;
 
 import java.util.List;
-
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import com.device_manager.model.Device;
 import com.device_manager.model.Employee;
 import com.device_manager.repository.EmployeePageableRepository;
-
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
 
@@ -22,6 +22,8 @@ public class EmployeeService {
 	private ObjectProvider<Employee> fakeEmployee;
 	
 	@Autowired EmployeePageableRepository employeeRepo;
+	
+	@Autowired DeviceService deviceService;
 	
 	public String createFakeEmployee() {
 		Employee e = fakeEmployee.getObject();
@@ -48,6 +50,9 @@ public class EmployeeService {
 	
 	public String deleteEmployee(Employee e) {
 		if (employeeRepo.existsById(e.getId())) {
+			if(e.getDevices().size() > 0) {
+				throw new DataIntegrityViolationException("Failed to delete employee because has relations with some devices!");
+			}
 			employeeRepo.delete(e);
 			return "Employee correctly removed from Database";
 		} else {
@@ -57,6 +62,9 @@ public class EmployeeService {
 	
 	public String deleteEmployee(Long id) {
 		if (employeeRepo.existsById(id)) {
+			if(employeeRepo.findById(id).get().getDevices().size() > 0) {
+				throw new DataIntegrityViolationException("Failed to delete employee because has relations with some devices!");
+			}
 			employeeRepo.deleteById(id);
 			return "Employee correctly removed from Database";
 		} else {
